refactor: :fire: join all entities per type
add Transfer, GoodAction, Benefit and CoinStatus entities in coin package because all is about coins
